Migrate from fractal to aiur, linting config + minor CSS changes
The changes for the CSS are as follows:

* the CSS for the sort arrows no longer uses asset files (with
  `faucet-pipeline-static`) but instead inlines the SVGs as a data URI
* certain properties for the `.tabelle-header` are now customizable via
  CSS Properties
* the CSS files have moved (because of the fractal -> aiur migration)

With the migration from fractal to aiur, a lot of the structure of the
repository has changed, but there should be no changes (apart from those
mentioned) in the assets which are generated from the CSS & JS. The aiur
config (similar to the fractal config before) is only really intended
for development and documentation and is not the end product of the
library.

This also changes the linting configuration -> tabs instead of spaces,
double quotes instead of single quotes, and enforcing semicolons.
In retrospect, this should have been it's own commit, but its a bit late
now.