Add filtering function to the Tabelle
For Tabelle we want a different kind of filtering than what is used by
default with List.js (e.g. we want && filtering, not || filtering).
To implement this, I've implemented custom search functionality using
the list object under the hood. I've also optimized the `values` of the
items in the list so that they use the `textContent` of the cell by
default which is then searched through. I think this is sufficient as a
default behavior -- it could be that we want to add a custom data
attribute in the future to customize the filter results, but it should
be ok for now.