Made name attribute for th header optional for tabelle-cljs
The `name` attribute is something that isn't really used for client-side
filtering (because you don't actually have to make any roundtrips to the
server). This makes the attribute optional, but repurposes it so that
you can use it to customize the internal CSS class that List.js will use
for sorting and filtering the column.