Added different implementation of sorting component
modified the implementation of the `.tabelle-arrows` component.
Inserting the arrows directly into the DOM is now deprecated: we
recommend wrapping them in a `<fieldset class="tabelle-arrows>`. All
changes should be backwards compatible, even though the component
behaves a bit differently now. However, if for some reason the
`.tabelle-arrow--asc` or `.tabelle-arrow--desc` labels are not wrapped
in a `.tabelle-header` class, this could cause bugs because the new
version uses absolute positioning to position the `::after` element of
the `.tabelle-arrow--asc` and `.tabelle-arrow--desc` in over the nearest
parent with `position: relative;`