Add focus styles and improve tab behavior within the header
This adds focus styles for the sort arrows (because it is actually the
label for the input field, we have to explicitly set the focus style on
the label). We use default focus styles which match default browser
styles, but we would expect that users will want to override this.