Update tabelle-cljs when changes to the HTML DOM occur
This adds a mutation-observer to the `tabelle-cljs` component which will
listen for new `tr` elements, `tr` elements that are removed, and any
changes to the HTML within a `tr` and update the index of the List.js
under the hood so that the element is also added/deleted/updated.

This is limited by the List.js implementation, so it is only possible to
delete a row element when the row is currently visible with the filters
which are currently set.